Decoupled object alife utils to simulation/alife.
