+++ src/engine/core/managers/simulation/utils/simulation_priority.ts
@@ -1,29 +1,7 @@
-import { registry } from "@/engine/core/database";
-import type { TSimulationObject } from "@/engine/core/managers/simulation";
-import type { Squad } from "@/engine/core/objects/server/squad";
-import { LuaLogger } from "@/engine/core/utils/logging";
-import { areObjectsOnSameLevel, getServerDistanceBetween } from "@/engine/core/utils/object/object_location";
-import { logicsConfig } from "@/engine/lib/configs/LogicsConfig";
-import { MAX_I32 } from "@/engine/lib/constants/memory";
+import { TSimulationObject } from "@/engine/core/managers/simulation";
+import { Squad } from "@/engine/core/objects/server/squad";
+import { areObjectsOnSameLevel, getServerDistanceBetween } from "@/engine/core/utils/object";
 import { Optional, ServerObject, TDistance, TRate } from "@/engine/lib/types";
-
-const logger: LuaLogger = new LuaLogger($filename);
-
-/**
- * Force updating all alife objects at once for instant and smooth alife spawn.
- */
-export function setUnlimitedAlifeObjectsUpdate(): void {
-  logger.info("Allow unlimited alife batched updates:", MAX_I32);
-  registry.simulator.set_objects_per_update(MAX_I32);
-}
-
-/**
- * Force updating stable count of alife objects.
- */
-export function setStableAlifeObjectsUpdate(): void {
-  logger.info("Set stable alife updating:", logicsConfig.ALIFE.OBJECTS_PER_UPDATE);
-  registry.simulator.set_objects_per_update(logicsConfig.ALIFE.OBJECTS_PER_UPDATE);
-}
 
 /**
  * Evaluates simulation priority by distance.
